Changed the load order making dotenv load first.
-Trying to solve problem with the production error. Heroku is giving an H10 error
which means the app crashed. Tests still pass in the development environment and
the error message `active_record/connection_adapters/connection_specification.rb:37:in `initialize': undefined method `tr' for nil:NilClass`
suggests that the problem is with the database connection. Dotenv gem should
only work in development but I'm troubleshooting to see if this works.

-Link about the dotenv load error
http://stackoverflow.com/questions/28953923/not-reading-env-variables-in-database-yml-rails-4-2-0-rubymine-7-postgres-ru